Task 10 completed

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,19 +8,19 @@
  */
 int _atoi(char *s)
 {
-	int i = 0;
-	int sign = 1;
-	int result = 0;
-	int started = 0;
+	long int i = 0;
+	long int sign = 1;
+	long int result = 0;
+	long int started = 0;
 
     /* Iterate through the string */
 	while (s[i] != '\0')
 	{
 	/* Handle '-' and '+' signs */
 		if (s[i] == '-')
-        {
+		{
 			sign = sign * -1;
-        }
+		}
 		else if (s[i] >= '0' && s[i] <= '9')
 		{
 			/* Convert character to integer */
